+++ resources/js/components/Admin/User/userComponent.vue
@@ -83,24 +83,27 @@
         },
         methods:{
             onDelete(id){
-                axios.delete('/api/deleteUser/'+id)
-                    .then(response => {
-                        this.$notify({
-                            title: 'Success',
-                            message: response.data.message,
-                            type: 'info'
-                        });
+                this.$confirm('Are you sure to delete this item?')
+                    .then(_ => {
+                        axios.delete('/api/deleteUser/' + id)
+                            .then(response => {
+                                this.$notify({
+                                    title: 'Success',
+                                    message: response.data.message,
+                                    type: 'info'
+                                });
 
-                    }).catch(error => {
-                    if (error.response) {
-                        this.$notify({
-                            title: 'Error',
-                            message: 'Error Input Data ',
-                            type: 'error'
+                            }).catch(error => {
+                            if (error.response) {
+                                this.$notify({
+                                    title: 'Error',
+                                    message: 'Error Input Data ',
+                                    type: 'error'
+                                });
+                            }
                         });
-                        /*this.errors = error.response.data.errors;*/
-                    }
-                });
+                    })
+                .catch(_ => {});
             }
         }
     }
